requirements.txt
# Python version requirement: 3.8+
# Core MCP Server dependencies
mcp>=1.0.0
asyncio-mqtt>=0.16.1
pydantic>=2.0.0,<3.0.0
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
duckduckgo-search>=8.1.1
aiohttp>=3.9.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
loguru>=0.7.2,<1.0.0
pydantic-settings>=2.0.0,<3.0.0
pyyaml>=6.0.0,<7.0.0 